WIP(integer-i64): Phase 2c — widen integer to i64, role-2 migration
Applies doc/claude/plans/01-integer-i64/2c.diff + introduces new 4-byte
raw accessors (get/set_u32_raw, get/set_i32_raw) to distinguish internal
collection state (vector/tree/hash/radix headers, File.ref, string-table
offsets) from user-level `integer` values.

Scope:
- src/store.rs: new raw accessors; get_int/set_int widen to i64;
  set_str/get_str/append_str use set_u32_raw for string length slots.
- src/vector.rs, tree.rs, hash.rs, radix_tree.rs: all role-2 header
  writes migrated to set_u32_raw / set_i32_raw.
- src/database/{allocation,io,structures,format,types}.rs: record-header
  + type-tag + string-offset writes migrated; read_data tp=0 splits from
  tp=6 (integer vs character).
- src/state/{io,text,mod}.rs, src/codegen_runtime.rs, src/native.rs,
  src/compile.rs, src/extensions.rs: File.ref sentinel checks switched
  to get/set_i32_raw; text-pos writes use set_u32_raw; reflection
  struct writes (StackFrame, ArgValue) use the right accessor per role.
- src/fill.rs: regenerated, then hand-patched for the unfixed i32/i64
  type-alias ambiguity (see "Known limitation" below).
- src/ops.rs: 2c.diff — ~18 op_*_int functions forward to op_*_long;
  dead _int macros removed; in-file tests updated to i64::MAX/MIN.
- src/state/codegen.rs: 7 × OpConstInt emission sites widen to i64.
- default/01_code.loft: integer size(4) → size(8).
- doc/claude/plans/01-integer-i64/: migration guide + 2c.diff updated.

Build: cargo build --release passes.  Tests not yet run.

Known limitation (not yet resolved):
The `pub type i32 = integer size(4)` alias cannot be distinguished from
the default `integer` type in the current Type system — both map to
`Type::Integer(i32::MIN+1, i32::MAX, false)`.  After 2c.diff widens
rust_type(Type::Integer) to "i64", every stdlib `const i32` operand
(e.g. OpCall's `d_nr`, `to`) generates `*s.code::<i64>()` — which reads
8 bytes from a bytecode slot that was designed for 4.  To make fill.rs
compile, `s.fn_call` was widened from `to: i32` to `to: i64` (and one
internal caller in state/mod.rs updated to match).

Follow-up needed for a stable Phase 2c:
1. Distinguish `i32` alias from default `integer` in Type — add size
   field to Type::Integer, or a narrower-range variant, so rust_type can
   return "i32" for the explicit alias.
2. Re-regen fill.rs once the type-system distinction is in place.
3. Audit remaining read_data / write_data fixed-offset reads for layout
   changes (PNG metadata, struct field size shifts from 4 → 8 byte
   integer fields, etc.).
4. Update test expectations (p180, p54_as_long_*, probe_00, overflow
   message panics).

+++ src/state/codegen.rs
@@ -182,7 +182,7 @@ impl State {
         match val {
             Value::Int(value) => {
                 stack.add_op("OpConstInt", self);
-                self.code_add(*value);
+                self.code_add(i64::from(*value));
                 I32.clone()
             }
             Value::Enum(value, tp) => {
@@ -411,8 +411,8 @@ impl State {
                 // Uses existing OpConstInt + OpVarRef — no new opcode needed.
                 // add_op → operator() already advances stack.position; no manual +4/+12 needed.
                 stack.add_op("OpConstInt", self);
-                self.code_add(*d_nr);
-                // clos_pos computed after ConstInt advanced stack.position by 4.
+                self.code_add(i64::from(*d_nr));
+                // clos_pos computed after ConstInt advanced stack.position by 8 (post-2c).
                 let clos_pos = stack.position - stack.function.stack(*clos_var);
                 stack.add_op("OpVarRef", self);
                 self.code_add(clos_pos);
@@ -723,7 +723,7 @@ impl State {
             match elem {
                 Type::Integer(_, _, _) | Type::Function(_, _, _) => {
                     stack.add_op("OpConstInt", self);
-                    self.code_add(0i32);
+                    self.code_add(0i64);
                 }
                 Type::Boolean => {
                     stack.add_op("OpConstFalse", self);
@@ -788,13 +788,13 @@ impl State {
                 stack.add_op("OpVarRef", self);
                 self.code_add(size_of::<crate::keys::DbRef>() as u16);
                 stack.add_op("OpConstInt", self);
-                self.code_add(0);
+                self.code_add(0i64);
                 stack.add_op("OpSetInt", self);
                 self.code_add(4u16);
                 stack.add_op("OpCreateStack", self);
                 self.code_add(size_of::<crate::keys::DbRef>() as u16);
                 stack.add_op("OpConstInt", self);
-                self.code_add(12);
+                self.code_add(12i64);
                 stack.add_op("OpSetByte", self);
                 self.code_add(4u16);
                 self.code_add(0u16);
@@ -822,7 +822,7 @@ impl State {
             }
             Type::Integer(_, _, _) | Type::Character => {
                 stack.add_op("OpConstInt", self);
-                self.code_add(i32::MIN);
+                self.code_add(i64::MIN);
             }
             Type::Long => {
                 stack.add_op("OpConstLong", self);
@@ -838,7 +838,7 @@ impl State {
             }
             Type::Boolean => {
                 stack.add_op("OpConstInt", self);
-                self.code_add(i32::MIN);
+                self.code_add(i64::MIN);
             }
             _ => {
                 // For other types, push a zero-filled DbRef as a generic null.
@@ -1135,7 +1135,7 @@ impl State {
                 // pre-init a fn-ref slot with 16 null bytes.
                 // d_nr = i32::MIN (integer null sentinel) + closure = null DbRef.
                 stack.add_op("OpConstInt", self);
-                self.code_add(i32::MIN);
+                self.code_add(i64::MIN);
                 stack.add_op("OpNullRefSentinel", self);
             } else {
                 // A5.6-1/A5.6-2: 16-byte fn-ref slot: [d_nr (4B)][closure DbRef (12B)].
